relocate some classes to remove more dependencies on other Zikula bundles. When deprecated code is fully removed, the dependencies on Zikula bundles will be gone
